+++ apps/ranger-next/next.svg.js
@@ -1,0 +1,6 @@
+module.exports = {
+  dir: 'svgs',
+  output: 'components',
+  componentName: 'SvgResource',
+  extension: 'ts'
+}