+++ packages/ranger-eslint-config/index.js
@@ -172,8 +172,8 @@ module.exports = {
     // 强制在关键字前后使用一致的空格 (前后腰需要)
     'keyword-spacing': 2,
     // 强制一行的最大长度
-    'max-len': [2, 200, { ignoreUrls: true }],
-    // 强制最大行数
+    'max-len': [2, 600, { ignoreUrls: true }],
+    // 强制最大行8
     'max-lines': 0,
     // 强制 function 定义中最多允许的参数数量
     'max-params': [1, 5],